Wiki Engine
In general terms, a wiki is a website where page content can easily be created, read, edited and removed through a front-end interface. Wikis usually contain multiple pages, mostly to do with what the underlying topic of the wiki is aimed at. Most wikis are navigated using a search function and links from other pages. Wikis can also be public or private; a public wiki means anyone can view or edit pages and a private wiki means only authorized users can view and change pages. A wiki engine is just the software/back-end that runs the wiki system. So in short, a wiki is a knowledge management system and a wiki engine is the software that runs the wiki.